[refactor] item 업로드 관련 함수를 WishItemRegistrationViewModel로 이동 및 WishItemViewModel에 item get/set 함수 추가
