+++ app/build.gradle
@@ -41,8 +41,8 @@ android {
         jvmTarget = "1.8"
     }
 
-    buildFeatures {
-        dataBinding = true
+    dataBinding {
+        enabled = true
     }
 }
