+++ app/src/main/java/com/hyeeyoung/wishboard/view/wish/item/screens/WishFragment.kt
@@ -14,14 +14,14 @@ import com.bumptech.glide.Glide
 import com.hyeeyoung.wishboard.R
 import com.hyeeyoung.wishboard.databinding.FragmentWishBinding
 import com.hyeeyoung.wishboard.util.extension.navigateSafe
-import com.hyeeyoung.wishboard.viewmodel.WishItemViewModel
+import com.hyeeyoung.wishboard.viewmodel.WishItemRegistrationViewModel
 import dagger.hilt.android.AndroidEntryPoint
 import kotlinx.coroutines.launch
 
 @AndroidEntryPoint
 class WishFragment : Fragment() { // TODO rename class name
     private lateinit var binding: FragmentWishBinding
-    private val viewModel: WishItemViewModel by hiltNavGraphViewModels(R.id.wish_item_registration_nav_graph)
+    private val viewModel: WishItemRegistrationViewModel by hiltNavGraphViewModels(R.id.wish_item_registration_nav_graph)
 
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
