+++ app/src/main/java/com/hyeeyoung/wishboard/viewmodel/WishListViewModel.kt
@@ -24,7 +24,6 @@ class WishListViewModel @Inject constructor(
     private val cartRepository: CartRepository,
 ) : ViewModel() {
     private var wishList = MutableLiveData<MutableList<WishItem>>(mutableListOf())
-    private var wishItem = MutableLiveData<WishItem>()
     private val wishListAdapter = WishListAdapter(application)
 
     private val token = prefs?.getUserToken()
@@ -76,12 +75,7 @@ class WishListViewModel @Inject constructor(
         }
     }
 
-    fun setWishItem(wishItem: WishItem) { // TODO WishItemViewModel로 이동
-        this.wishItem.value = wishItem
-    }
-
     fun getWishList(): LiveData<MutableList<WishItem>?> = wishList
-    fun getWishItem(): LiveData<WishItem> = wishItem
     fun getWishListAdapter(): WishListAdapter = wishListAdapter
 
     companion object {
