+++ app/src/main/res/layout/activity_wish_link_sharing.xml
@@ -8,7 +8,7 @@
 
         <variable
             name="viewModel"
-            type="com.hyeeyoung.wishboard.viewmodel.WishItemViewModel" />
+            type="com.hyeeyoung.wishboard.viewmodel.WishItemRegistrationViewModel" />
     </data>
 
     <androidx.constraintlayout.widget.ConstraintLayout
