+++ app/src/main/res/layout/fragment_wish.xml
@@ -8,7 +8,7 @@
 
         <variable
             name="viewModel"
-            type="com.hyeeyoung.wishboard.viewmodel.WishItemViewModel" />
+            type="com.hyeeyoung.wishboard.viewmodel.WishItemRegistrationViewModel" />
 
         <import type="android.view.View" />
     </data>
