+++ app/src/main/res/layout/fragment_wish_item_detail.xml
@@ -8,7 +8,7 @@
 
         <variable
             name="viewModel"
-            type="com.hyeeyoung.wishboard.viewmodel.WishListViewModel" />
+            type="com.hyeeyoung.wishboard.viewmodel.WishItemViewModel" />
 
         <import type="androidx.navigation.Navigation" />
